:recycle: Refactor GameRoom codes
GameRoom 페이지와 관련된 코드 리팩토링
